+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v17.2.2.6 / 2017 Aug 28
+* **Fix** - Regression in Loggly support that was preventing logs from being searchable
+* **Add** - Added ability to add runtime tags to `Logger` via `Logger.AddTag()`
+* **Add** - Support for tagging in Elassticsearch
+ 
+```csharp
+[GuaranteedRate.Sextant "17.2.2.6"]
+```
+
 ## v17.2.2.5 / 2017 Aug 22
 * **Add** - Added ability to add global runtime tags for `StatsDMetrics`